Castor3D: Removed specific PassBuffers.
