+++ source/Core/Castor3D/Castor3DPch.cpp
@@ -1,3 +1,1 @@
 #include "Castor3DPch.hpp"
-
-#include "Castor3D/Castor3DPrerequisites.hpp"
